Atualização 0.2: - Classes renomeadas.
- Consertado: erro que o retorno "jogo iniciado com sucesso" aparecia por último.

- Modificado: retorno de mensagem. De madeira do tipo: "carvalho", para "bruta".
